fix: handle 'recipes' module not found error
Lesson: 5. Authentication 06:02 min

While running the Django migration command, an
error popped up. It couldn't find the 'recipes'
module, which should be in the list of installed
apps. This update will fix this error by making
sure Django settings can find and import the
'recipes' module correctly.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	modified:   main/settings.py
	modified:   main/urls.py
	modified:   templates/includes/footer.html
	modified:   templates/includes/header.html